requirements.txt
backcall==0.2.0
cfgv==3.3.1
cloudpickle==2.0.0
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.4
entrypoints==0.3
filelock==3.4.2
fonttools==4.28.5
identify==2.4.2
ipykernel==6.6.1
ipython==7.31.0
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
nodeenv==1.6.0
numpy==1.22.0
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
platformdirs==2.4.1
pre-commit==2.16.0
prompt-toolkit==3.0.24
ptyprocess==0.7.0
Pygments==2.11.2
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
spyder-kernels==2.2.0
threadpoolctl==3.0.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
virtualenv==20.13.0
wcwidth==0.2.5
wurlitzer==3.0.2
